Move UbloxFirmware7 to its own files.
